requirements.txt
Flask==2.3.2
PyMuPDF==1.21.1
nltk==3.8
google-generativeai==0.3.2